syncval: Segregate CommandBufferAccessContext to file
