+++ BUILD.gn
@@ -239,6 +239,8 @@ vvl_sources = [
   "layers/stateless/stateless_validation.h",
   "layers/sync/sync_access_context.cpp",
   "layers/sync/sync_access_context.h",
+  "layers/sync/sync_commandbuffer.cpp",
+  "layers/sync/sync_commandbuffer.h",
   "layers/sync/sync_common.cpp",
   "layers/sync/sync_common.h",
   "layers/sync/sync_model.cpp",
