+++ layers/sync/sync_sync_op.cpp
@@ -17,6 +17,7 @@
 
 #include "sync/sync_sync_op.h"
 #include "sync/sync_renderpass.h"
+#include "sync/sync_commandbuffer.h"
 
 // Range generators for to allow event scope filtration to be limited to the top of the resource access traversal pipeline
 //
